Kreirani interfejsi i implementiran dependency injection.

+++ PopsiTopsi/PresentationLayer/ArticleForm.cs
@@ -1,5 +1,7 @@
 ﻿using BusinessLayer;
-using DataLayer.Models;
+using Shared.BusinessInterface;
+using Shared.Models;
+
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -14,14 +16,18 @@ namespace PresentationLayer
 {
     public partial class ArticleForm : Form
     {
-        public readonly ArticleBusiness articleBusiness;
+        public readonly IArticleBusiness articleBusiness;
+        public readonly IStuffBusiness stuffBusiness;
         public OpenFileDialog dialog;
         public Image img;
         public byte[] byteimg;
-        public ArticleForm()
+   
+        public ArticleForm(IArticleBusiness _articleBusiness, IStuffBusiness _stuffBusiness)
         {
             InitializeComponent();
-            this.articleBusiness = new ArticleBusiness();
+            articleBusiness = _articleBusiness;
+            stuffBusiness = _stuffBusiness;
+          
         }
         public void UpdateDataGrid()
         {
@@ -56,23 +62,28 @@ namespace PresentationLayer
 
         private void btnArticalAdd_Click(object sender, EventArgs e)
         {
-            string type;
+            if (PresentationLayer.Properties.Settings.Default.UserId == 1003)
+            {           
+             string type;
             List<String> list = new List<String>();
-           foreach (CheckBox cb in panel1.Controls.OfType<CheckBox>())
+            foreach (CheckBox cb in panel1.Controls.OfType<CheckBox>())
             {
-                if(cb.Checked)
+                if (cb.Checked)
                 {
                     list.Add(cb.Text);
                 }
             }
-            type = String.Join(",",list.ToArray());
+            type = String.Join(",", list.ToArray());
+            try { 
             Article article = new Article();
             article.Name = tbArticalName.Text;
             article.Price = Decimal.Parse(tbArticalPrice.Text);
             article.Type = type;
             article.Description = tbArticalDescription.Text;
             article.Image = Convert.ToBase64String(byteimg);
-
+            
+                
+                
             bool result = this.articleBusiness.InsertArticle(article);
             if (result)
                 MessageBox.Show("Uspesan unos artikla!");
@@ -80,7 +91,17 @@ namespace PresentationLayer
                 MessageBox.Show("Unos artikala nije uspesan!");
             UpdateDataGrid();
             DeleteTextBox();
-           
+            }
+                catch
+                {
+                    MessageBox.Show("Unosite validne podatke!");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Morate se prijaviti kao admin da bi dodavali artikle!");
+                DeleteTextBox();
+            }
         }
         Image ConvertBinaryToImage(byte[] data)
         {
@@ -107,12 +128,12 @@ namespace PresentationLayer
                     lblFileName.Text = openFile.FileName;
                     img = Image.FromFile(lblFileName.Text);
                     byteimg = ConvertImageToBinary(img);
-                    pictureBox1.Image =img ;
-                    
+                    pictureBox1.Image = img;
+
                 }
             }
         }
-        
+
         private void dgArticles_Cell(object sender, DataGridViewCellEventArgs e)
         {
             int idSelect = Convert.ToInt32(dgArticles.SelectedRows[0].Cells["Id"].Value.ToString());
@@ -122,29 +143,62 @@ namespace PresentationLayer
             tbArticalPrice.Text = article.Price.ToString();
             tbArticalDescription.Text = article.Description;
             byteimg = Convert.FromBase64String(article.Image);
-            pictureBox1.Image = ConvertBinaryToImage(Convert.FromBase64String(article.Image));          
+            pictureBox1.Image = ConvertBinaryToImage(Convert.FromBase64String(article.Image));
+            if (article.Type.Contains("POPSI"))
+                cbPops.Checked = true;
+            else
+                cbPops.Checked = false;
+            if (article.Type.Contains("KROFNE"))
+                cbDonut.Checked = true;
+            else
+                cbDonut.Checked = false;
+            if (article.Type.Contains("MAFINI"))
+                cbMuffin.Checked = true;
+            else cbMuffin.Checked = false;
+            if (article.Type.Contains("BOX"))
+                cbBox.Checked = true;
+            else
+                cbBox.Checked = false;
         }
 
         private void btnArticalDelete_Click(object sender, EventArgs e)
         {
-            int idSelect = Convert.ToInt32(dgArticles.SelectedRows[0].Cells["Id"].Value.ToString());
-           
-            bool result = this.articleBusiness.DeleteArticle(idSelect);
-            if(result)
+            if (PresentationLayer.Properties.Settings.Default.UserId == 1003)
             {
-                MessageBox.Show("Uspesno brisanje!");
+               
+                int idSelect = Convert.ToInt32(dgArticles.SelectedRows[0].Cells["Id"].Value.ToString());
+                 if(articleBusiness.IsArticleConnectedToBill(idSelect)==true)
+                {
+                    MessageBox.Show("Ne mozete izbrisati ovu stavku jer se nalazi na racunu!");
+                }
+                 else
+                {
+
+                
+                bool result = this.articleBusiness.DeleteArticle(idSelect);
+                if (result)
+                {
+                    MessageBox.Show("Uspesno brisanje!");
+                }
+                else
+                {
+                    MessageBox.Show("Neuspesno brisanje!");
+                }
+                UpdateDataGrid();
+                DeleteTextBox();
+                }
             }
             else
             {
-                MessageBox.Show("Neuspesno brisanje!");
+                MessageBox.Show("Morate se prijaviti kao admin da bi brisali!");
             }
-            UpdateDataGrid();
-            DeleteTextBox();
         }
 
         private void btnArticalChange_Click(object sender, EventArgs e)
         {
-             string type;
+            if (PresentationLayer.Properties.Settings.Default.UserId == 1003)
+            { 
+                string type;
             List<String> list = new List<String>();
             foreach (CheckBox cb in panel1.Controls.OfType<CheckBox>())
             {
@@ -162,7 +216,7 @@ namespace PresentationLayer
             article.Description = tbArticalDescription.Text;
             article.Image = Convert.ToBase64String(byteimg);
             bool result = articleBusiness.UpdateArticle(article, idSelect);
-            if(result)
+            if (result)
             {
                 MessageBox.Show("Uspesna izmena!");
             }
@@ -172,6 +226,12 @@ namespace PresentationLayer
             }
             UpdateDataGrid();
             DeleteTextBox();
+        }
+            else
+            {
+                MessageBox.Show("Morate se prijaviti kao admin da bi menjali artikle!");
+            }
+        
         }
     }
 }
